requirements.txt
alembic==1.0.11
attrs==19.1.0
awesome-slugify==1.6.5
bcrypt==3.1.7
beautifulsoup4==4.7.1
cryptography==2.8
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
regex==2019.6.8
requests==2.22.0
requests-mock==1.6.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
urllib3==1.25.3
user-agent==0.1.9
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.5.2
